Use patched ToHashDocument, to fix bug
Fog::ToHashDocument has a bug which appears to result in the
attributes of an XML entity that also contains a text node to be stripped.

In our case, this makes:

<Connection ns12:ipAddress="192.0.2.1">Test</Connection>

become:

:Connection => Test

I've copied ToHashDocument into our own parser (OurToHashDocument) for now, and
patched to fix the issue.

Needs discussion though.

Now it returns

:Connection => {
   attributes => { ns12:ipAddress => "192.0.2.1" },
   value => "Test",
}

... which is up for discussion.