Add mock for a vAppTemplate
This just returns a single record, which is the intention of a look-up by vAppTemplate name and Catalog name. The fields are correct, but the values are bogus. We are very specific about how it is called because that is what we require for our test – this could be extended to deal with other complex queries.