CGI.escape and character replacements
`URI.escape` is used here to URL-encode the URL sent to the vCloud
Director API in the `get_execute_query` method:

https://github.com/fog/fog/blob/v1.29.0/lib/fog/vcloud_director/requests/compute/get_execute_query.rb#L96

`URI.escape` does not URL-encode `+` characters, meaning that vCloud
Director (we're using version 5.5 update 2) interprets the `+` as a
space.

This causes the following query to fail:

    type=task&sortDesc=startDate&filter=name==networkConfigureEdgeGatewayServices;objectName==mygateway;startDate=ge=2015-04-09T11:13:54.373+01:00

...because vCloud Director interprets the date stamp as:

    2015-04-09T11:13:54.373 01:00

Note the missing `+` between the time and the timezone.

`URI.escape` is [deprecated][] as of Ruby 1.9.2, so we looked at using
`CGI.escape` instead, which seems to be aimed at encoding query
parameter values rather than a whole URL as `URI.escape` was. See this
discussion on the `ruby-core` mailing list:

http://blade.nagaokaut.ac.jp/cgi-bin/vframe.rb/ruby/ruby-core/29293?29179-31097

We then encountered a further problem in that by URL-encoding the
parameter value, the double equals (`==`) used to separate sub-parameter
keys from their values, i.e.
`name==networkConfigureEdgeGatewayServices`, was also being URL-encoded,
which the vCloud Director API rejects.

Looking at the RFCs for URLs, it seems this behaviour (using an `=`
inside a parameter value) is valid:

http://blog.lunatech.com/2009/02/03/what-every-web-developer-must-know-about-url-encoding#Thereservedcharactersaredifferentforeachpart
https://tools.ietf.org/html/rfc3986#section-2.4

...so the proper solution is to URL-encode each sub-parameter value
individually. This will require us to modify the API for the
`get_execute_query` method.

We have not been able to find a way around by using `URI.escape` or `CGI.escape`
that fixes this issue other than running a replacement of specific
characters.